refactor: 금액 관련한 객체 추상화
