+++ src/main/java/christmas/domain/common/Amount.java
@@ -2,14 +2,27 @@ package christmas.domain.common;
 
 public abstract class Amount {
 
-    private final int amount;
+    private static final int MIN_AMOUNT = 0;
 
-    protected Amount(int amount) {
+    private final long amount;
+
+    protected Amount(long amount) {
+        validateMinAmount(amount);
         this.amount = amount;
     }
 
-    public boolean isGreaterOrEqualThan(int amount) {
+    private static void validateMinAmount(long amount) {
+        if (amount < MIN_AMOUNT) {
+            throw new IllegalArgumentException("amount %d should be greater or equal than 0.".formatted(amount));
+        }
+    }
+
+    public boolean isGreaterOrEqualThan(long amount) {
         return this.amount >= amount;
+    }
+
+    public long getAmount() {
+        return amount;
     }
 
 }
