+++ src/main/java/christmas/domain/event/benefit/TotalBenefitAmount.java
@@ -1,0 +1,11 @@
+package christmas.domain.event.benefit;
+
+import christmas.domain.common.Amount;
+
+public class TotalBenefitAmount extends Amount {
+
+    public TotalBenefitAmount(int amount) {
+        super(amount);
+    }
+
+}
