+++ src/main/java/christmas/domain/event/discount/WeekdaysDiscountPolicy.java
@@ -26,9 +26,9 @@ public class WeekdaysDiscountPolicy implements DiscountPolicy {
     @Override
     public DiscountAmount discount() {
         if (meetsDayRangeCriterion()) {
-            return DiscountAmount.from(calculateDiscountAmount());
+            return new DiscountAmount(calculateDiscountAmount());
         }
-        return DiscountAmount.from(0);
+        return new DiscountAmount(0);
     }
 
     private boolean meetsDayRangeCriterion() {
